+++ AGENTS.md
@@ -141,3 +141,4 @@ When making significant changes:
 - tower-lsp 0.20 defaults to `buffer_unordered(4)` for all messages. Without `.concurrency_level(1)`, `did_change` notifications can be processed out of order, corrupting incremental text sync.
 - When propagating function-scoped packages across `source()` edges, treat nested source-call scopes as descendants of outer package scopes; exact interval equality breaks nested local `source()` inheritance.
 - When adding a new VS Code-exposed LSP setting, update all three extension layers together: `editors/vscode/package.json` schema contribution, `editors/vscode/src/extension.ts` initialization-options forwarding, and `editors/vscode/src/test/settings.test.ts` so docs and Settings UI do not drift from server support.
+- Avoid duplicating initialization-options builders or types between the VS Code runtime and tests; share a single factory/module so settings like cache forwarding and master defaults (for example `diagnostics.enabled`) cannot drift.
